+++ src/main/java/christmas/model/benefit/WeekendDiscount.java
@@ -25,7 +25,7 @@ public class WeekendDiscount extends Benefit {
     }
 
     @Override
-    public int getDiscountAmount(Plan plan) {
+    public long getDiscountAmount(Plan plan) {
         return calculateDiscountPerMenu(plan.getMainMenuOrders(), discountAmountPerMenu);
     }
 }
